Move location of file containing PaniniEngine
