+++ lib/engine.js
